Image_Classification.ipynb
This project is an image classification model using TensorFlow and Keras to identify fruits and vegetables. It loads training, validation, and test datasets from user-defined directories and resizes all images to a uniform shape. A Convolutional Neural Network (CNN) is built, trained, and evaluated for accuracy and loss over several epochs. The model can then predict the category of a new image and display the result with confidence. Finally, the trained model is saved for future use or deployment.